[WFLY-456][WFLY-490] Refactor audit log handler test to not use platform mbean server
